feat: add localization to header
