+++ src/components/Header/i18n/de.json
@@ -1,0 +1,5 @@
+{
+  "login": "einloggen",
+  "logout": "ausloggen",
+  "signup": "anmelden"
+}
