+++ src/components/Header/i18n/en.json
@@ -1,0 +1,5 @@
+{
+  "login": "log in",
+  "logout": "log out",
+  "signup": "sign up"
+}
