+++ src/components/Header/i18n/ru.json
@@ -1,0 +1,5 @@
+{
+  "login": "войти",
+  "logout": "выйти",
+  "signup": "регистрация"
+}
