+++ src/components/Header/index.ts
@@ -1,0 +1,3 @@
+import Header from './HeaderContainer';
+
+export default Header;
